web/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==14.1
python-multipart==0.0.20
mcp==1.1.0
anthropic==0.40.0
python-dotenv==1.0.1
